peer review feedback
RENAMED: controller to fix spelling

Refactored MyRandom
refactored graphic tools to another file
started to re-factor to with a GameEntity design
GameEntity: Turrets
GameEntity: SpaceShip
GameEntity: Lerper
ADDED:TODO.txt
ADDED: more designs
ADDED: Generating a heart.xlsx
ADDED: ParticalEffects

git-svn-id: https://svn.neumont.edu:8443/svn/cplusplus_anthony_corbin@56 c70e0d21-08ce-ef44-a38d-91c21494a941

+++ GameSolution/Engine/MyRandom.cpp
@@ -1,21 +1,22 @@
 #include "MyRandom.h"
 #include <random>
-#include <ctime>
+#include <ctime>//for srand seed
 #include "Vector 2.h"
-int Random::randomInt(int low, int high) {
+
+int      Random::randomInt(int low, int high) {
 	std::random_device rd;//dont like making multiple engines
 	std::mt19937 mt(rd()); //seed
 	std::uniform_int_distribution<int>dist(low,high);
 	return dist(mt);
 }
-float Random::randomFloat(float low, float high) {
+float    Random::randomFloat(float low, float high) {
 	return randomFloat() * (high-low) + low;
 }
-void Random::init() {
+void     Random::init() {
 	srand((unsigned)time(0));
 }
-float Random::randomFloat() {
-	return rand()/(static_cast<double>(RAND_MAX));
+float    Random::randomFloat() {
+	return rand()/(static_cast<float>(RAND_MAX));
 }
 Vector2D Random::randomUnitVector() {
 	return Vector2D(randomFloat(-1,1),randomFloat(-1,1)).normalized();
